MODCLUSTER-575 Move DeterministicLoadState out of the SPI to package org.jboss.modcluster.load.metric.impl
