+++ core/src/main/java/org/jboss/modcluster/load/metric/impl/DeterministicLoadState.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.modcluster.load.metric;
+package org.jboss.modcluster.load.metric.impl;
 
 /**
  * @author Paul Ferraro
